testing/web-platform/tests/web-animations/interfaces/KeyframeEffect/effect-easing.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Effect-level easing tests</title>
<link rel="help" href="http://w3c.github.io/web-animations/#calculating-the-transformed-time">
<link rel="author" title="Hiroyuki Ikezoe" href="mailto:hiikezoe@mozilla-japan.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<script src="../../resources/easing-tests.js"></script>
<body>
<div id="log"></div>
<div id="target"></div>
<script>
"use strict";

gInvalidEasings.forEach(function(invalidEasing) {
  test(function(t) {
    var div = createDiv(t);
    assert_throws({ name: 'TypeError' },
                  function() {
                    div.animate({ easing: invalidEasing }, 100 * MS_PER_SEC);
                  });
  }, 'Invalid keyframe easing value: \'' + invalidEasing + '\'');
});

</script>
</body>
